refactor(order-status): rely on forms status
