+++ tests/unit/order-status-lookup.test.js
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import { test } from 'node:test';
-import deriveOrderStatusKey from '../../widgets/forms/order-status-state.js';
+import { deriveOrderStatusKey } from '../../widgets/forms/order-status-lookup.js';
 
 test('uses the normalized status returned by the forms action', () => {
   assert.equal(
@@ -9,26 +9,8 @@ test('uses the normalized status returned by the forms action', () => {
   );
 });
 
-test('derives processed for a booked EBS order without deliveries', () => {
-  assert.equal(
-    deriveOrderStatusKey({ succeeded: true, order: { state: 'Booked' } }),
-    'processed',
-  );
-});
-
-test('maps EBS line-item statuses like Magento', () => {
-  assert.equal(
-    deriveOrderStatusKey({ succeeded: true, order: { lineItem: [{ status: 'Booked', quantity: '1' }] } }),
-    'processed',
-  );
-  assert.equal(
-    deriveOrderStatusKey({ succeeded: true, order: { lineItem: [{ status: 'Entered', quantity: '1' }] } }),
-    'received',
-  );
-  assert.equal(
-    deriveOrderStatusKey({ succeeded: true, order: { lineItem: [{ status: 'Closed', quantity: '1' }] } }),
-    'shipped',
-  );
+test('returns unavailable for unsuccessful responses', () => {
+  assert.equal(deriveOrderStatusKey({ succeeded: false }), 'unavailable');
 });
 
 test('keeps delivery-based shipped detection as a fallback', () => {
