+++ widgets/forms/order-status-lookup.js
@@ -1,7 +1,4 @@
 import { getLocaleAndLanguage, getFormSubmissionUrl } from '../../scripts/scripts.js';
-import deriveOrderStatusKey from './order-status-state.js';
-
-export { default as deriveOrderStatusKey } from './order-status-state.js';
 
 /**
  * Loads localized order-status copy from the sibling `order-status.json`.
@@ -16,6 +13,31 @@ export async function loadOrderStatusCopy(lang) {
   const data = await resp.json();
   const key = data[lang] ? lang : 'en';
   return data[key];
+}
+
+/**
+ * Derives the display status from the Forms API response.
+ *
+ * The Forms API calculates the authoritative Magento-compatible status and
+ * returns it as order.status. Delivery data remains a fallback while older
+ * Forms deployments are still in use.
+ *
+ * @param {Record<string, any>|null} result - Parsed API response
+ * @returns {string} Status key matching a key in the localized copy
+ */
+export function deriveOrderStatusKey(result) {
+  if (!result?.succeeded) return 'unavailable';
+
+  const normalizedStatus = result.order?.status;
+  if (['received', 'processed', 'partiallyShipped', 'shipped', 'cancelled', 'unavailable'].includes(normalizedStatus)) {
+    return normalizedStatus;
+  }
+
+  const deliveries = [].concat(result.order?.delivery ?? []);
+  const shippedCount = deliveries.filter((d) => d.shipped).length;
+  if (shippedCount === 0) return deliveries.length ? 'processed' : 'received';
+  if (shippedCount < deliveries.length) return 'partiallyShipped';
+  return 'shipped';
 }
 
 /**
